host should mute on system wake
these days, muted is a sane default.
it is easy enough to notice when my system
is wrongly muted, and unmute. but accidental
sounds are somewhat unbecoming.

One downside to this is that lingering tabs or
other noisemakers could be hidden. it would be
worth figuring a way to show a visual bell every
so often if sounds are played under mute.